requirements.txt
anthropic
google-generativeai
openai
python-dotenv
ipython